# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Platform API
  description: |
    Comprehensive Healthcare Platform API for managing patients, providers, 
    appointments, claims, payments, and more.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limits
    - Anonymous endpoints: 30 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Admin endpoints: 30 requests/minute
    
    ## Versioning
    API version is specified in the URL path (e.g., `/api/v1/...`) or 
    via the `X-API-Version` header.
  version: 1.0.0
  contact:
    name: Healthcare Platform Support
    email: support@healthcare.com
    url: https://healthcare.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.healthcare.com/v1
    description: Production server
  - url: https://staging-api.healthcare.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /providers/search:
    get:
      tags: [Providers]
      summary: Search healthcare providers
      description: Search and filter healthcare providers by various criteria
      operationId: searchProviders
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query (name, specialty, condition)
          schema:
            type: string
        - name: specialty
          in: query
          description: Filter by specialty
          schema:
            type: string
        - name: city
          in: query
          description: Filter by city
          schema:
            type: string
        - name: state
          in: query
          description: Filter by state
          schema:
            type: string
        - name: insurance
          in: query
          description: Filter by accepted insurance
          schema:
            type: string
        - name: min_rating
          in: query
          description: Minimum rating (1-5)
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: telehealth_available
          in: query
          description: Filter for telehealth availability
          schema:
            type: boolean
        - name: accepting_new_patients
          in: query
          description: Filter for providers accepting new patients
          schema:
            type: boolean
        - name: latitude
          in: query
          description: User latitude for distance calculation
          schema:
            type: number
        - name: longitude
          in: query
          description: User longitude for distance calculation
          schema:
            type: number
        - name: max_distance
          in: query
          description: Maximum distance in miles
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [relevance, rating, distance, experience, reviews]
            default: relevance
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderSearchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /payments:
    get:
      tags: [Payments]
      summary: List payments
      operationId: listPayments
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'

    post:
      tags: [Payments]
      summary: Process a payment
      operationId: processPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

  # ==================== Reports ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external integrations
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
...
    Provider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        professional_title:
          type: string
        specialties:
          type: array
          items:
            type: string
        primary_specialty:
          type: string
        bio:
          type: string
        rating:
          type: number
        average_rating:
          type: number
        total_reviews:
          type: integer
        years_of_experience:
          type: integer
        languages_spoken:
          type: array
          items:
            type: string
        gender:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        address_line1:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        distance:
          type: number
          description: Distance in miles (if user coordinates provided)
        accepts_new_patients:
          type: boolean
        telehealth_available:
          type: boolean
        consultation_fee:
          type: number
        insurance_accepted:
          type: array
          items:
            type: string
        hospital_affiliations:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            type: string
        board_certifications:
          type: array
          items:
            type: string
        profile_image_url:
          type: string
        next_available:
          type: string
          format: date
    ProviderSearchResponse:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/Provider'
        pagination:
          $ref: '#/components/schemas/Pagination'
        filters:
          type: object
          properties:
            applied:
              type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        patient_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: USD
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        payment_method:
          type: string
        transaction_id:
          type: string
        created_at:
          type: string
          format: date-time
    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProcessPaymentRequest:
      type: object
      required: [patient_id, amount, payment_method]
      properties:
        patient_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: USD
        payment_method:
          type: string
        appointment_id:
          type: integer
        metadata:
          type: object
...